+++ src/components/notice/index.tsx
@@ -6,8 +6,8 @@ export { default as TimeInformation } from "./timeInfro/TimeInformation";
 
 export { default as NoticeBtn } from "./button/NoticeBtn";
 
-export { default as NoticeButtonSection } from "./section/NoticeButtonSection";
+export { default as NoticeButtonSection } from "./button/NoticeButtonSection";
 
-export { default as NoticeKeywordButtonSection } from "./section/NoticeKeywordButtonSection";
+export { default as NoticeKeywordButtonSection } from "./button/NoticeKeywordButtonSection";
 
 export { default as NoticeKeywordInputSection } from "./section/NoticeKeywordInputSection";
